Fix compare of vertex Z and test Z in gSPBranchLessZ.
Fixed GT64 - Championship Edition: missing polygon #151

+++ src/gSP.h
@@ -148,7 +148,7 @@ void gSPDMAVertex( u32 v, u32 n, u32 v0 );
 void gSPCBFDVertex( u32 v, u32 n, u32 v0 );
 void gSPDisplayList( u32 dl );
 void gSPBranchList( u32 dl );
-void gSPBranchLessZ( u32 branchdl, u32 vtx, f32 zval );
+void gSPBranchLessZ( u32 branchdl, u32 vtx, u32 zval );
 void gSPDlistCount(u32 count, u32 v);
 void gSPSprite2DBase(u32 _base );
 void gSPDMATriangles( u32 tris, u32 n );
